cms-api: Refactor User module to build
