--- cms-api/src/modules/user/user.interface.ts
@@ -1,9 +1,0 @@
-export interface IUser {
-    username: string;
-    email: string;
-    password: string;
-    role: any[];
-
-    created: Date;
-    changed: Date;
-}
